onDrawDebug: angleScale float not double (fmuls + single fctiwz) (92.77 -> 93.30 fn, unit -> 96.262)

+++ src/partyobj.cpp
@@ -5449,7 +5449,7 @@ void CGPartyObj::onDrawDebug(CFont* font, float x, float& y, float z)
 		}
 
 		float rate = static_cast<float>(CharaGhostValue(0x2054)) / kMonObjPercentMax;
-		double angleScale;
+		float angleScale;
 		switch (bossKind) {
 		default:
 			angleScale = kMonObjOne;
